Add node_modules/, remove keys.js
Remember to always add node_modules/ to your .gitignore.

The node_modules/ is expectedly large and not something
you want taking up space/bandwith on the repos.

With npm and a package.json, we can pass a lightweight
application around a team and download the most up-to-
date packages from NPM.